+++ barelymusician/dsp/gain_processor.cpp
@@ -48,7 +48,8 @@ double ApplyLinearRamp(double gain, double target_gain, int ramp_frame_count, do
 }  // namespace
 
 GainProcessor::GainProcessor(int frame_rate) noexcept
-    : unity_ramp_frame_count_(FramesFromSeconds(frame_rate, kUnityRampDurationSeconds)) {}
+    : unity_ramp_frame_count_(
+          static_cast<double>(FramesFromSeconds(frame_rate, kUnityRampDurationSeconds))) {}
 
 void GainProcessor::Process(double* buffer, int channel_count, int frame_count) noexcept {
   int frame = 0;
